+++ sdk/startup/watch.go
@@ -38,7 +38,7 @@ func Watch(waitTimeBeforeExit time.Duration) {
 
 	cid, _ := crypto.RandomString(8)
 	if cid == "" {
-		cid = "VSECMSDK"
+		panic("Unable to create a secure correlation id.")
 	}
 
 	for {
